flats controller + new/create functions and page
