+++ app/controllers/flats_controller.rb
@@ -1,0 +1,21 @@
+class FlatsController < ApplicationController
+  def new
+    @flat = Flat.new
+  end
+
+  def create
+    @flat = Flat.new(flat_params)
+    @flat.user = current_user
+    if @flat.save!
+      redirect_to flat_path(@flat)
+    else
+      render :new
+    end
+  end
+
+  private
+
+  def flat_params
+    params.require(:flat).permit(:name, :rooms, :price)
+  end
+end
