+++ app/controllers/pages_controller.rb
@@ -2,23 +2,4 @@ class PagesController < ApplicationController
   def home
     @flats = Flat.all
   end
-
-  def new
-    @flat = Flat.new
-  end
-
-  def create
-    @flat = Flat.new(flat_params)
-    if @flat.save!
-      redirect_to flat_path(@flat)
-    else
-      render :new
-    end
-  end
-
-  private
-
-  def flat_params
-    params.require(:flat).permit(:name, :rooms, :price)
-  end
 end
