+++ app/models/flat.rb
@@ -1,3 +1,4 @@
 class Flat < ApplicationRecord
   belongs_to :user
+
 end
