+++ app/views/flats/new.html.erb
@@ -1,13 +1,14 @@
 <div class="container">
-  <h1 class="text-center mt-4">Add new list</h1>
+  <h1 class="text-center mt-4">List your property</h1>
   <div class="row justify-content-center">
     <div class="col-12 col-md-6 col-lg-4">
      <%= simple_form_for(@flat) do |f| %>
         <%= f.input :name %>
         <%= f.input :rooms %>
-        <%= f.input :price %> €
+        <%= f.input :price %>
         <%= f.submit class: "btn btn-primary" %>
       <% end %>
+      <%= link_to "Cancel", flats_path, class: "btn btn-danger my-2" %>
     </div>
   </div>
 </div>
